Added Basic admin console
You can now download or generate the zip file from it.
I used for the first time the @wraps command, will go and change
current login checker to it.
I also added a salt to the HASHIDs generater, which is defined in
config. I should change config to a .gitignored file. Also
complete a .gitignore