Added config file and gitignored it
As the config is important and contains secret keys etc, I don't
want to upload that information to github when I put it up. As
such that stuff is now in 'application.cfg'.

Contents of 'application.cfg':

SECRET_KEY = 'development key'
HASHID_KEY = 'this is a key'
ADMIN_UNAME = 'admin'
ADMIN_PWORD = 'pword'
DEBUG = True

(These are all development ones and will not be used in production)