Refactored decorators, added @async, changed email
I added an async decorator to run things in the back ground with
it's used for sending emails and generating the .zip file, so that
there is a zippy (ha) response to the client.

With the addition of this decorator I decided it was about time to
do a small amount of refactoring, by placing all of the decorators
into it's own file - decorators.py.

I also added a bit of code in the send_email function, so that it
will only send to the test email if debug is true. While testing
this however I ran across a few issues where it seems like gmail
would send the email and it would work fine on gmail email
addresses but not on various temporary addresses i tested. This
is gonna require a bit more testing, especailly on puppycat