Changed template to allow markers no house
In the database markers without a house have always been allowed however when
Lana was implementing the design for school she made it so that it would
have to have a house otherwise the template would fail. We were tired and didn'
have much time so we made the decision to just force it. We did come across one
time during the week where we wanted to add a neutral marker for our school's
principle, so I have implemented this.

Changes still need to be made to the gimp script I used but that is not part of
this repo. Note that on the svg file it will be put down as 'None' for the
house.