Added more data options for houses
In the database houses now have the extra attributes, 'shortdesc' a shortened
description that can be used on things like marker and user pages, and 'color'.
'color' is a hex code (without the starting hash) that is currently only used
on the graph page but may infuture be added to some of the css as the highlight
color for the house's markers etc.

Also added a bunch of comments.