Changed admin panel to run on Flask-admin
This is a fairly decent change but it does not change user
functionality at all. I've implemented the flask-admin library
to provide crud functionality to Houses, Users and Markers.

I've used SQLALCHEMY and a new line in application.cfg is required
defining SQLALCHEMY_DATABASE_URI but I may change this later.

Currently a large portion of the code is still using pymysql but
I will transfer that over soon. I still need to add some batch
editing features (markers being available and the like), as well
as more information on the index page, and updates to the login
system, and possibly marker generation but this is a nice big step.