Added images to the houses in database
In the original run of IGNITE we did what could be considered a slight
hack. Instead of actually uploading images we chucked them with the
statics and gave them the file name housename.jpg. It worked but
not really the best idea.

To fix this I have placed an extra column in the houses table,
'imagepath' and used flask-admin to build an upload form.

Images upload to 'static/upload'.

I have also used the house colors on their pages for their title.