Added the static pages into database
Originally the way to add a static page was to add it as a template
and then add a line of code for it in __init__.py however this is
not ideal for less technical people. To fix this I have added a
pages table to the database which can be accessed from the flask
admin page. When it's added the links are automatically generated
to it. It can choose to inherit from the original layout or not.

I have also changed the house links up the top to display from the
database rather than their values being hardcoded.